fix(v0.18.0): Decimal migration & trading settlement locking (B9-B10)
- B9: migrate all float money fields to Decimal with Numeric(20, 8)
  columns across trading (trading.py, inter_chain.py, offer_sync),
  marketplace (main.py, matching_service, marketplace_service),
  wallet bridge (bridge_db.py, price_api.py, bridge_routes.py);
  add Alembic migration for trading app; update routers and services
  to use Decimal for price/amount parameters
- B10: add SELECT FOR UPDATE to settlement.py and matching_engine.py
  to prevent race conditions; implement idempotency checks for
  lock-escrow and settle endpoints; add amount > 0 validation to
  inter_chain trade creation; add idempotency key support to
  exchange_compat payment creation

+++ apps/wallet/src/wallet_app/bridge/bridge_routes.py
@@ -3,6 +3,7 @@ ETH-AIT Bridge API Routes
 REST API endpoints for bridge operations.
 """
 
+from decimal import Decimal
 from typing import Annotated, Any
 
 from fastapi import APIRouter, Depends, HTTPException
@@ -115,7 +116,7 @@ async def complete_deposit(
 
 
 @router.get("/calculate")
-async def calculate_exchange(eth_amount: float) -> dict[str, Any]:
+async def calculate_exchange(eth_amount: Decimal) -> dict[str, Any]:
     """
     Calculate AIT amount for a given ETH amount.
